+++ .eslintrc.json
@@ -25,10 +25,6 @@
             "error",
             4
         ],
-        /*"linebreak-style": [
-            "error",
-            "windows"
-        ],*/
         "quotes": [
             "error",
             "double"
