+++ .gitignore
@@ -10,7 +10,6 @@
 
 # production
 /build
-
 # misc
 .DS_Store
 .env.local
@@ -18,6 +17,7 @@
 .env.test.local
 .env.production.local
 .idea
+.build
 
 npm-debug.log*
 yarn-debug.log*
