+++ src/components/OverviewPage.tsx
@@ -3,12 +3,12 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router";
-import { fetchSubordinates, setFetching, setSubordinatesList } from "../../../redux/ducks/subordinates";
+import { fetchSubordinates, setFetching, setSubordinatesList } from "../redux/ducks/subordinates";
 
-import OverviewPageStyled from "../../../styled/mainPage/overviewPage/OverviewPageStyled";
+import OverviewPageStyled from "../styled/mainPage/overviewPage/OverviewPageStyled";
 import { Col, Container, Row } from "react-bootstrap";
-import Preloader from "../../../assets/images/preloader.svg";
-import {AppDispatch, RootState} from "../../../redux/store";
+import Preloader from "../assets/images/preloader.svg";
+import {AppDispatch, RootState} from "../redux/store";
 
 type ParamType = {
     name:string
@@ -50,9 +50,7 @@ const OverviewPage:React.FC = ():JSX.Element => {
                     <Col xl="12" lg="12" md="12" sm="12" xs="12">
                         <OverviewPageStyled.MyUL>
                             {(employerDataSelector[1])?(
-                            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                            // @ts-ignore
-                                employerDataSelector[1]["direct-subordinates"].map(
+                                Object.values(employerDataSelector[1]).map(
                                     ( subordinate: string , index: React.Key ) =>
                                         <OverviewPageStyled.MyLI key={index}>
                                             {subordinate}
@@ -60,10 +58,11 @@ const OverviewPage:React.FC = ():JSX.Element => {
                                 !isFetchingSelector&&<div>Haven`t subordinates</div>
                             }
                             {
-                                isFetchingSelector&&<div>
-                                    <OverviewPageStyled.MyIMG src = {Preloader} />
-                                </div>
-                            }
+                                isFetchingSelector&&(
+                                    <div>
+                                        <OverviewPageStyled.MyIMG src = {Preloader} />
+                                    </div>
+                                )}
                         </OverviewPageStyled.MyUL>
                     </Col>
                 </Row>
