+++ src/components/index.ts
@@ -1,4 +1,4 @@
-import MainPage from "./mainPage/MainPage";
-import OverviewPage from "./mainPage/overviewPage/OverviewPage";
+import MainPage from "./MainPage";
+import OverviewPage from "./OverviewPage";
 
 export default { MainPage, OverviewPage };
